refactor: Separate context provider.
Create a new 'DarkModeContextProvider' components and
separate the 'DarkModeContext' code from the 'App.js' file.
