+++ src/components/navbar/NavBar.jsx
@@ -2,7 +2,7 @@ import React, { useContext } from "react";
 import "./Navbar.module.css";
 import navbarStyle from "./Navbar.module.css";
 import { Link, useLocation } from "react-router-dom";
-import { DarkModeContext } from "../../App";
+import { DarkModeContext } from "../../contexts/DarkModeContext";
 
 const tabList = [
   { title: "TO DON'T LIST", path: "/todontlist" },
