+++ src/contexts/DarkModeContext.jsx
@@ -1,0 +1,37 @@
+import React, { createContext, useEffect, useState } from "react";
+import { getLocalStoage, setLocalStoage } from "../utils/localStorage";
+
+export const DarkModeContext = createContext();
+
+export function DarkModeProvider({ children }) {
+  const [darkMode, setDarkMode] = useState(false);
+
+  useEffect(() => {
+    const isDarkMode =
+      getLocalStoage("theme") === "dark" ||
+      (!("theme" in localStorage) &&
+        window.matchMedia("(prefers-color-scheme: dark)").matches);
+    setDarkMode(isDarkMode);
+    updateDarkMode(isDarkMode);
+  }, []);
+
+  const toggleDarkMode = () => {
+    setDarkMode(!darkMode);
+    updateDarkMode(!darkMode);
+  };
+  return (
+    <DarkModeContext.Provider value={{ darkMode, toggleDarkMode }}>
+      {children}
+    </DarkModeContext.Provider>
+  );
+}
+
+function updateDarkMode(darkMode) {
+  if (darkMode) {
+    document.documentElement.classList.add("dark");
+    setLocalStoage("theme", "dark");
+  } else {
+    document.documentElement.classList.remove("dark");
+    setLocalStoage("theme", "light");
+  }
+}
